+++ tests/unit/lbrynet_daemon/test_ExchangeRateManager.py
@@ -11,14 +11,13 @@ from tests import util
 class FeeFormatTest(unittest.TestCase):
     def test_fee_created_with_correct_inputs(self):
         fee_dict = {
-            'USD': {
-                'amount': 10.0,
-                'address': "bRcHraa8bYJZL7vkh5sNmGwPDERFUjGPP9"
-            }
+            'currency':'USD',
+            'amount': 10.0,
+            'address': "bRcHraa8bYJZL7vkh5sNmGwPDERFUjGPP9"
         }
         fee = Fee.FeeValidator(fee_dict)
-        self.assertEqual(10.0, fee['USD']['amount'])
-
+        self.assertEqual(10.0, fee['amount'])
+        self.assertEqual('USD', fee['currency'])
 
 class FeeTest(unittest.TestCase):
     def setUp(self):
@@ -26,11 +25,11 @@ class FeeTest(unittest.TestCase):
 
     def test_fee_converts_to_lbc(self):
         fee_dict = {
-            'USD': {
-                'amount': 10.0,
-                'address': "bRcHraa8bYJZL7vkh5sNmGwPDERFUjGPP9"
+            'currency':'USD',
+            'amount': 10.0,
+            'address': "bRcHraa8bYJZL7vkh5sNmGwPDERFUjGPP9"
             }
-        }
+
         rates = {
             'BTCLBC': {'spot': 3.0, 'ts': util.DEFAULT_ISO_TIME + 1},
             'USDBTC': {'spot': 2.0, 'ts': util.DEFAULT_ISO_TIME + 2}